pyproject.toml
[tool.poetry]
name = "autodialer"
version = "0.1.0"
description = ""
authors = ["SirAbhi13 <abhinav.sny.2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.8"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
requests = "^2.32.3"
twilio = "^9.2.3"
celery = "^5.4.0"
django-celery-results = "^2.5.1"
sqlalchemy = "^2.0.32"
django-extensions = "^3.2.3"
ipython = "^8.26.0"
django-filter = "^24.3"
phonenumbers = "^8.13.43"
pytz = "^2024.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
isort = "^5.13.2"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
